sentinelx/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
python-dotenv
langchain-core
langchain-google-genai
langgraph
requests
